requirements.txt
tqdm~=4.66.5
python-dotenv~=1.0.1
loguru~=0.7.2
transformers~=4.44.2
FlagEmbedding~=1.2.11
openai~=1.47.1
pyserini~=0.38.0
faiss-cpu
numpy~=1.26.4
scikit-learn~=1.5.2
scipy
tiktoken~=0.7.0
pydantic~=2.9.2
pandas~=2.2.3
torch~=2.4.1
uvicorn~=0.30.6
fastapi~=0.115.0
sentence-transformers~=3.1.1
cohere~=5.11.3
jieba~=0.42.1
rank-bm25~=0.2.2